pyproject.toml
[tool.poetry]
name = "basic-django"
version = "0.1.0"
description = ""
authors = ["Rasul Kireev <rasul@builtwithdjango.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-allauth = "^0.50.0"
python-webpack-boilerplate = "^1.0.0"
django-widget-tweaks = "^1.4.12"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
pylint = "^2.17.1"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.djlint]
profile="django"
ignore = "H031"
